"Fix" compile warnings

+++ src/ConvertTools.hpp
@@ -9,7 +9,7 @@ public:
     static int MH_HexToInt(std::vector<unsigned char> hexValArr) {
         int num = 0;
         
-        for(int i = 0; i < hexValArr.size(); i++) {
+        for(size_t i = 0; i < hexValArr.size(); i++) {
             num |= hexValArr[i] << i * 8;
         }
 
@@ -19,7 +19,7 @@ public:
     static float MH_HexToFloat(std::vector<char> hexValArr) {
         char tempArr[sizeof(float)];
 
-        for(int i = 0; i < hexValArr.size(); i++) {
+        for(size_t i = 0; i < hexValArr.size(); i++) {
             tempArr[i] = hexValArr[i];
         }
 
@@ -29,7 +29,7 @@ public:
     static double MH_HexToDouble(std::vector<char> hexValArr) {
         char tempArr[sizeof(double)];
 
-        for(int i = 0; i < hexValArr.size(); i++) {
+        for(size_t i = 0; i < hexValArr.size(); i++) {
             tempArr[i] = hexValArr[i];
         }
